Migrate ComplexInterface test cases to JSR223
